+++ src/settings.example.py
@@ -375,6 +375,7 @@ sensors = {
             'pin_pdsck': 'P21',
             'scale': 4.424242,
             'offset': -73000,
+            'decimals': 3,
         },
         {
             'id': 'ds18b20-1',